+++ CriticalMass/SettingsTableSectionHeader.swift
@@ -1,0 +1,13 @@
+//
+//  SettingsTableSectionHeader.swift
+//  CriticalMaps
+//
+//  Created by Malte Bünz on 19.04.19.
+//  Copyright © 2019 Pokus Labs. All rights reserved.
+//
+
+import UIKit
+
+class SettingsTableSectionHeader: UITableViewHeaderFooterView, NibProviding {
+    @IBOutlet var titleLabel: UILabel!
+}