+++ CriticalMass/SettingsViewController.swift
@@ -8,58 +8,15 @@
 import UIKit
 
 class SettingsViewController: UITableViewController {
-    enum Section: CaseIterable {
-        case preferences
-        case github
-        case info
+    private let themeController: ThemeController!
 
-        struct Model {
-            var title: String?
-            var action: Action
-        }
+    init(themeController: ThemeController) {
+        self.themeController = themeController
+        super.init(nibName: nil, bundle: nil)
+    }
 
-        var numberOfRows: Int {
-            return models.count
-        }
-
-        var secionTitle: String? {
-            switch self {
-            case .preferences,
-                 .github:
-                return nil
-            case .info:
-                return NSLocalizedString("settings.section.info", comment: "")
-            }
-        }
-
-        var cellClass: UITableViewCell.Type {
-            switch self {
-            case .preferences:
-                return SettingsSwitchTableViewCell.self
-            case .github:
-                return SettingsGithubTableViewCellTableViewCell.self
-            case .info:
-                return SettingsInfoTableViewCell.self
-            }
-        }
-
-        var models: [Model] {
-            switch self {
-            case .preferences:
-                return [Model(title: NSLocalizedString("GPS", comment: ""), action: .none)]
-            case .github:
-                return [Model(title: nil, action: .open(url: URL(string: "https://github.com/criticalmaps/criticalmaps-ios")!))]
-            case .info:
-                return [Model(title: NSLocalizedString("settings.website", comment: ""), action: .open(url: URL(string: "https://www.criticalmaps.net")!)),
-                        Model(title: NSLocalizedString("settings.twitter", comment: ""), action: .open(url: URL(string: "https://twitter.com/criticalmaps/")!)),
-                        Model(title: NSLocalizedString("settings.facebook", comment: ""), action: .open(url: URL(string: "https://www.facebook.com/criticalmaps")!))]
-            }
-        }
-
-        enum Action {
-            case open(url: URL)
-            case none
-        }
+    required init?(coder _: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
 
     override func viewDidLoad() {
@@ -73,6 +30,8 @@ class SettingsViewController: UITableViewController {
 
         configureSettingsFooter()
         configureNavigationBar()
+
+        tableView.register(SettingsTableSectionHeader.nib, forHeaderFooterViewReuseIdentifier: SettingsTableSectionHeader.typeName)
     }
 
     override func viewDidLayoutSubviews() {
@@ -97,12 +56,34 @@ class SettingsViewController: UITableViewController {
         }
     }
 
+    // MARK: Actions
+
+    @IBAction func gpsCellAction(_ sender: UISwitch) {
+        Preferences.gpsEnabled = sender.isOn
+    }
+
+    @IBAction func darkModeCellAction(_ sender: UISwitch) {
+        let theme: Theme = sender.isOn ? .dark : .light
+        themeController.changeTheme(to: theme)
+        themeController.applyTheme()
+    }
+
     override func numberOfSections(in _: UITableView) -> Int {
         return Section.allCases.count
     }
 
-    override func tableView(_: UITableView, titleForHeaderInSection section: Int) -> String? {
-        return Section.allCases[section].secionTitle
+    override func tableView(_: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        let cell = tableView.dequeueReusableHeaderFooterView(withIdentifier: SettingsTableSectionHeader.typeName)
+        let header = cell as! SettingsTableSectionHeader
+        header.titleLabel.text = Section.allCases[section].secionTitle
+        return cell
+    }
+
+    override func tableView(_: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        guard section == Section.info.index else {
+            return 0.0
+        }
+        return 42.0
     }
 
     override func tableView(_: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -113,19 +94,9 @@ class SettingsViewController: UITableViewController {
         return 60
     }
 
-    override func tableView(_: UITableView, willDisplayHeaderView view: UIView, forSection _: Int) {
-        for subView in view.subviews {
-            subView.backgroundColor = .white
-            for case let label as UILabel in subView.subviews {
-                label.font = UIFont.systemFont(ofSize: 24, weight: .bold)
-            }
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    override func tableView(_: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let section = Section.allCases[indexPath.section]
-        let identifier = String(describing: section.cellClass)
-        let cell = tableView.dequeueReusableCell(withIdentifier: identifier, for: indexPath)
+        let cell = settingsCell(for: section, at: indexPath)
         cell.textLabel?.text = section.models[indexPath.row].title
         return cell
     }
@@ -149,15 +120,36 @@ class SettingsViewController: UITableViewController {
     }
 }
 
-extension UITableViewController {
-    func sizeFooterToFit() {
-        guard let footerView = tableView.tableFooterView else {
-            return
+extension SettingsViewController {
+    fileprivate func settingsCell(for section: Section, at indexPath: IndexPath) -> UITableViewCell {
+        var cell: UITableViewCell
+        switch section {
+        case .preferences:
+            guard let switchCell = tableView.dequeueReusableCell(withIdentifier: String(describing: section.cellClass), for: indexPath) as? SettingsSwitchTableViewCell else {
+                fatalError("Should be a SettingsSwitchCell")
+            }
+            let model = section.models[indexPath.row]
+            if model.title == String.gpsLocalizedString {
+                let isGPSEnabled = Preferences.gpsEnabled
+                let gpsHandler: SettingsSwitchHandler = { [unowned self] settingsSwitch in
+                    self.gpsCellAction(settingsSwitch)
+                }
+                switchCell.configure(isOn: isGPSEnabled, handler: gpsHandler)
+                cell = switchCell
+            } else if model.title == String.themeLocalizedString {
+                let isNightModeEnabled = themeController.currentTheme == .dark ? true : false
+                let nightModeHandler: (UISwitch) -> Void = { [unowned self] settingsSwitch in
+                    self.darkModeCellAction(settingsSwitch)
+                }
+                switchCell.configure(isOn: isNightModeEnabled, handler: nightModeHandler)
+                cell = switchCell
+            } else {
+                print("Title not found")
+            }
+            cell = switchCell
+        case .info, .github:
+            cell = tableView.dequeueReusableCell(withIdentifier: String(describing: section.cellClass), for: indexPath)
         }
-        let height = footerView.systemLayoutSizeFitting(UIView.layoutFittingCompressedSize).height
-        let footerFrame = footerView.frame
-        if height != footerFrame.size.height {
-            footerView.frame.size.height = height
-        }
+        return cell
     }
 }